Highlight update and new menu functionality
I've done a bit more work on the highlighting function, now each element stays highlighted as long as the mouse is over it. However this has caused another issue (properly explained in code comments) that has to be fixed at some point.
I have also added another function tied to menu button 3 which essentially just scrolls the page to the position of the graph when clicked. This is a very general function, and could quite easily be modified so that the page scrolls to any particular HTML element in the html.index file