Added new functions for previewing graphs
The functionality changes are all contained in the svgDiagrams.js file. Essentially, a modal containing a picture and caption unique to each svg element will now pop up when the menu button labelled 'preview' is clicked. Further more, the toInfo function has been re-used to create the toGraph function, which essentially does the same thing, except it now jumps to images.
The index.html file has had some images added to it to demonstrate the new functions, while the styles.css file has had code added to it that governs how the modal is displayed on the screen, as well as its sub-elements.